CoinBot/requirements.txt
python-binance
pandas
ta
python-dotenv